+++ test/interpreter/function-isnontext.spec.ts
@@ -32,14 +32,14 @@ describe('Function ISNONTEXT', () => {
     expect(engine.getCellValue(adr('B1'))).toEqualError(detailedError(ErrorType.NA, ErrorMessage.WrongArgNumber))
   })
 
-  // Inconsistency with Product 1
   it('range value results in VALUE error', () => {
     const engine = HyperFormula.buildFromArray([
       ['=4/1'],
-      ['=4/0', '=ISNONTEXT(A1:A3)'],
+      ['=4/0'],
       ['=4/2'],
+      ['=ISNONTEXT(A1:A3)'],
     ])
 
-    expect(engine.getCellValue(adr('B2'))).toEqualError(detailedError(ErrorType.VALUE, ErrorMessage.WrongType))
+    expect(engine.getCellValue(adr('A4'))).toEqualError(detailedError(ErrorType.VALUE, ErrorMessage.WrongType))
   })
 })